Move WPComAccountCard.
