+++ js/src/setup-mc/setup-stepper/setup-accounts/wordpressdotcom-account/wpcom-account-card/index.js
